+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Calculator</title>
     <link rel="stylesheet" type="text/css" href="src/css/index.css" />
-    <script src="src/js/index.js"></script>
+    <script type="module" src="src/js/index.js"></script>
   </head>
   <body>
     <div id="app">
