+++ src/js/Calculator.js
@@ -1,7 +1,8 @@
 class Calculator {
   num1 = 0;
   num2 = 0;
-  total = 0;
+  op = null;
+  current = 0;
 
   add(n, m) {
     return n + m;
@@ -14,6 +15,13 @@ class Calculator {
   }
   divide(n, m) {
     return Math.floor(n / m);
+  }
+
+  initialize() {
+    this.num1 = 0;
+    this.num2 = 0;
+    this.op = null;
+    this.current = 0;
   }
 }
